+++ src/main/kotlin/org/wfanet/measurement/duchy/deploy/common/postgres/readers/ComputationStageAttemptReader.kt
@@ -19,6 +19,7 @@ import org.wfanet.measurement.common.db.r2dbc.ReadContext
 import org.wfanet.measurement.common.db.r2dbc.boundStatement
 import org.wfanet.measurement.internal.duchy.ComputationStageAttemptDetails
 
+/** Performs read operations on Computations and ComputationStageAttempts tables */
 class ComputationStageAttemptReader {
 
   /**
